자유게시판 - 드래그앤드롭 첨부파일  - 게시글 삭제  1. 게시글 show = n으로 변경  2. 로컬에 저장된 첨부파일 삭제(AJAX와 Array 사용)  3. DB에 저장된 첨부파일 삭제

+++ paiksrecipe/src/main/java/com/paiksrecipe/controller/AjaxUploadController.java
@@ -15,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -94,32 +95,41 @@ public class AjaxUploadController {
 	}
 	
 	@ResponseBody
-	@PostMapping("/upload/deleteFile")
-	public ResponseEntity<String> deleteFile(String fileName) {
-		log.info("fileName : " + fileName);
-		// fileName = /2020/04/10/s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
-		// 확장자 검사
-		String formatName = fileName.substring(fileName.lastIndexOf(".") + 1);
-		// formatName = jpg
-		MediaType mType = MediaUtils.getMediaType(formatName);
-		if(mType != null) { // 이미지 파일이면 원본이미지 삭제
-			String front = fileName.substring(0, 12);
-			// front = /2020/04/10/
-			String end = fileName.substring(14);
-			// end = ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
-			// File.separatorChar : 유닉스 / 윈도우즈 \
-			new File(uploadPath + (front + end).replace('/', File.separatorChar)).delete();
-			// new File(c://developer/upload + /2020/04/10/ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
-			// replace >> c:\\developer\ upload + \2020\04\10\ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
-			// delete >> c:\\developer\ upload + \2020\04\10\ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
-			// 원본이미지만 삭제
+	@PostMapping("/upload/deleteAllFile")
+	public ResponseEntity<String> deleteFile(@RequestParam("files[]") String[] files) {
+		log.info("delete all files : " + files);
+
+		if(files == null || files.length == 0) {
+			return new ResponseEntity<String>("deleted", HttpStatus.OK);
 		}
-		// 원본 파일 삭제(이미지이면 썸네일 삭제)
-		new File(uploadPath + fileName.replace('/', File.separatorChar)).delete();
-		// new File(c://developer/upload/2020/04/10/s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
-		// replace >> c:\\developer\ upload\2020\04\10\s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
-		// delete >> c:\\developer\ upload\2020\04\10\s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
 		
+		for(String fileName : files) {
+			
+			// fileName = /2020/04/10/s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
+			// 확장자 검사
+			String formatName = fileName.substring(fileName.lastIndexOf(".") + 1);
+			// formatName = jpg
+			MediaType mType = MediaUtils.getMediaType(formatName);
+
+			if(mType != null) { // 이미지 파일이면 원본이미지 삭제
+				String front = fileName.substring(0, 12);
+				// front = /2020/04/10/
+				String end = fileName.substring(14);
+				// end = ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
+				// File.separatorChar : 유닉스 / 윈도우즈 \
+				new File(uploadPath + (front + end).replace('/', File.separatorChar)).delete();
+				// new File(c://developer/upload + /2020/04/10/ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
+				// replace >> c:\\developer\ upload + \2020\04\10\ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
+				// delete >> c:\\developer\ upload + \2020\04\10\ + ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png
+				// 원본이미지만 삭제
+			}
+			
+			// 원본 파일 삭제(이미지이면 썸네일 삭제)
+			new File(uploadPath + fileName.replace('/', File.separatorChar)).delete();
+			// new File(c://developer/upload/2020/04/10/s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
+			// replace >> c:\\developer\ upload\2020\04\10\s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
+			// delete >> c:\\developer\ upload\2020\04\10\s_ca6412fe-c76a-469f-87c4-c8fb032c39de_005_004.png)
+		}
 		// ResponseEntity : request를 처리하고 response할 때 response설정을 미세하게 변경할 때 사용
 		return new ResponseEntity<String>("deleted", HttpStatus.OK);
 	}
